2008-09-17 Roman - /DB/MySQL: For UTF codepages length multiplication to 2 for string is made. - FIX:/UI/QTCfg: Save on exit is fixed. - FIX:/SYS: Fixed DB fields set is fixed for TArchives, TBDS and TTransports. - FIX:/SYS/Elements: Short DB fields length atribute is fixed.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@743 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/telem.h
@@ -112,7 +112,7 @@ class TFld
 	//Attributes
 	string		m_name;		// Name of element (name column into BD);
 	string		m_descr;	// Description of element;
-	short int	m_len;		// field len
+	int		m_len;		// field len
 	short int	m_dec	:5;	// field dec (for real)
 	short int	m_type	:4;	// Type (Dec, Hex, Oct, ...)
 	unsigned	m_flg;		// element flags (Selected, SelfFld ...);
